requirements.txt
# Web
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1

# Core ML (CPU wheels)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
torchvision==0.17.1+cpu
torchaudio==2.2.1+cpu

# Model + utils
transformers==4.43.2
tokenizers==0.19.1
efficientnet_pytorch==0.7.1
Pillow==10.3.0
numpy==1.26.4
tqdm==4.66.4

# HF hub to pull private model
huggingface_hub==0.23.2
